requirements.txt
playwright==1.49.0
PyYAML==6.0.2
